requirements.txt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos
beautifulsoup4
easyocr
numpy
opencv-python
python-decouple
requests
selenium
tqdm
requests_toolbelt
fastapi
uvicorn
python-multipart